Add CMake, add access specifiers.
