+++ exor.h
@@ -1,7 +1,8 @@
 #ifndef EXOR_H
 #define EXOR_H
 
-#include "systemc.h"
+#include <systemc.h>
+
 #include "nand.h"
 
 // TODO: Insert your code here
