+++ mon.h
@@ -1,25 +1,26 @@
 #ifndef MONITOR_H
 #define MONITOR_H
 
-#include "systemc.h"
 #include <iostream>
-using namespace std;
+#include <systemc.h>
 
 SC_MODULE(Monitor)
 {
+public:
     sc_in<bool> A, B, Z;
-
-    void monitor()
-    {
-        cout << sc_time_stamp()  << "\t" << A << "\t" << B << "\t" << Z <<endl;
-    }
 
     SC_CTOR(Monitor)
     {
-        cout << endl <<  "time\tA\tB\tF" << endl;
+        std::cout << std::endl <<  "time\tA\tB\tF" << std::endl;
         SC_METHOD(monitor);
         sensitive << A << B << Z;
         dont_initialize();
+    }
+
+private:
+    void monitor()
+    {
+        std::cout << sc_time_stamp()  << "\t" << A << "\t" << B << "\t" << Z << std::endl;
     }
 };
