+++ nand.h
@@ -1,7 +1,8 @@
 #ifndef NAND_H
 #define NAND_H
 
-#include "systemc.h"
+#include <systemc.h>
+
 
 // TODO: Insert your code here
